Make the API key the last argument
At all endpoints, move the API key so that it is the last argument of
the generated function.

Some sort of consistency is obviously necessary: either the API key
should always be the first argument, or it should always be the last
argument.  If we made the API type nested rather than flat, we would
clearly want the key to be the outermost path element (perhaps just
under "v1"), but we're going with a flat design.  Since I intend to
provide a `MonadReader` version of the API that reads the API key from
the environment, it seems right to put the key at the end.