Switch from `POSIXTime` to `UTCTime`
- With the aid of `generic-data-surgery`, we can easily modify incoming
data.  We now convert incoming POSIX timestamps to UTC time, because
that's just better.
- The one exception: we still use POSIX timestamps when sending requests
for download links.  This is because the expiry time is given by Nexus
Mods as a POSIX timestamp, and it's easier for the end-user if we don't
make them convert.
- This commit also adds a few modules that make the surgery easier.
Generic surgery operations are easily expressed as operations in an
indexed state monad, so we define those in `NexusMods.Surgery`.  We
enable `QualifiedDo` for `do`-notation in indexed monads, and we put the
aliases in `NexusMods.Indexed`.
- Also, we now use surgery to get a few other `FromJSON` instances.
- We need to enable a few extra extensions in order to express the
types in `NexusMods.Surgery`.